rag_scorecard_project/requirements.txt
# Core
torch>=2.3.0
transformers>=4.41.0
accelerate>=0.29.3
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
protobuf>=3.19,<5.0

# RAG orchestration
llama-index>=0.11.32
llama-index-retrievers-bm25>=0.5.2
llama-index-vector-stores-faiss>=0.1.0
llama-index-llms-huggingface>=0.1.4
langchain>=0.2.2
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Ingestion & embeddings
google-genai>=1.4.0
llama-index-embeddings-google-genai>=0.1.0
llama-index-embeddings-huggingface>=0.5.3

# Evaluation
ragas>=0.1.7
deepeval>=0.4.5
pandas>=2.2.2
jsonlines>=4.0.0
